fix: i18n support custom config
