+++ packages/smallfish-plugin-i18n/src/i18n.ts
@@ -1,22 +1,6 @@
 import i18next from 'i18next';
-import languageDetector from 'i18next-browser-languagedetector';
+import i18nextLanguageDetector from 'i18next-browser-languagedetector';
 
-import resources from '@/i18n';
-
-Object.keys(resources).map(key => {
-  resources[key] = {
-    translation: resources[key],
-  };
-});
-
-const params = {
-  resources,
-  fallbackLng: {
-    en: ['en-Us'],
-    default: ['en-US'],
-  },
-};
-
-i18next.use(languageDetector).init(params);
+export const languageDetector = i18nextLanguageDetector;
 
 export default i18next;
